requirements.txt
fastapi
uvicorn
groq
gtts
python-dotenv
opencv-python
mediapipe
